Added all different imports (not yet completly working)
