+++ madcad/rendering/d3/mesh.py
@@ -1,12 +1,15 @@
 """Displays for mesh data:  cloud points, webs, meshes, ... """
 
+from . import settings
+import numpy.core as np
+from .rendering.base import Display
 
 class MeshDisplay(Display): # WARNING: Old `SolidDisplay` class
     ''' Display render Meshes '''
     def __init__(self, scene, positions, normals, faces, lines, idents, color=None):
         self.box = npboundingbox(positions)
         self.options = scene.options
-        
+
         s = settings.display
         color = fvec3(color or s['solid_color'])
         line = (    (length(s['line_color']) + dot(color-s['solid_color'], s['line_color']-s['solid_color']))
@@ -25,33 +28,38 @@ class MeshDisplay(Display): # WARNING: Old `SolidDisplay` class
             wire.append((f[1], f[2]))
             wire.append((f[2], f[0]))
         self.disp_wire = LinesDisplay(scene, self.vertices, wire, color=line, alpha=0.3, layer=-1e-6)
-        
+
     def stack(self, scene):
         yield ((), 'screen', -1, self.vertices.prerender)
-        if self.options['display_faces']:    
+        if self.options['display_faces']:
             yield ((), 'screen', 0, self.disp_faces.render)
             yield ((), 'ident', 0, self.disp_faces.identify)
-        else:                                
+        else:
             yield ((), 'screen', 1, self.disp_ghost.render)
             yield ((), 'ident', 0, self.disp_ghost.identify)
-        if self.options['display_groups']:    yield ((), 'screen', 1, self.disp_groups.render)
-        if self.options['display_points']:    yield ((), 'screen', 2, self.disp_points.render)
-        if self.options['display_wire']:    yield ((), 'screen', 2, self.disp_wire.render)
-    
-    @property
-    def world(self):    return self.vertices.world
-    @world.setter
-    def world(self, value):    self.vertices.world = value
+        if self.options['display_groups']:
+            yield ((), 'screen', 1, self.disp_groups.render)
+        if self.options['display_points']:
+            yield ((), 'screen', 2, self.disp_points.render)
+        if self.options['display_wire']:
+            yield ((), 'screen', 2, self.disp_wire.render)
 
-    #def control(self, view, key, sub, evt):
-        #if evt.type() == QEvent.MouseButtonRelease and evt.button() == Qt.LeftButton:
-            #sub = sub[0]
-            #flags, idents = self.vertices.flags, self.vertices.idents
-            #for i in range(len(idents)):
-                #flags[i] ^= idents[i] == sub
-            #self.vertices.flags_updated = True
-            #view.update()
-            #evt.accept()
+    @property
+    def world(self):
+        return self.vertices.world
+    @world.setter
+    def world(self, value): 
+        self.vertices.world = value
+
+    # def control(self, view, key, sub, evt):
+    #     if evt.type() == QEvent.MouseButtonRelease and evt.button() == Qt.LeftButton:
+    #         sub = sub[0]
+    #         flags, idents = self.vertices.flags, self.vertices.idents
+    #         for i in range(len(idents)):
+    #             flags[i] ^= idents[i] == sub
+    #         self.vertices.flags_updated = True
+    #         view.update()
+    #         evt.accept()
 
 
 class WebDisplay(Display):
@@ -64,29 +72,32 @@ class WebDisplay(Display):
         self.disp_edges = LinesDisplay(scene, self.vertices, lines, color=color, alpha=1, layer=-2e-6)
         self.disp_groups = PointsDisplay(scene, self.vertices, points, layer=-3e-6)
         self.disp_points = PointsDisplay(scene, self.vertices, range(len(positions)), layer=-1e-6)
-        
 
     def stack(self, scene):
         yield ((), 'screen', -1, self.vertices.prerender)
-        if self.options['display_groups']:        yield ((), 'screen', 2, self.disp_groups.render)
-        if self.options['display_points']:        yield ((), 'screen', 2, self.disp_points.render)
+        if self.options['display_groups']:
+            yield ((), 'screen', 2, self.disp_groups.render)
+        if self.options['display_points']:
+            yield ((), 'screen', 2, self.disp_points.render)
         yield ((), 'screen', 1, self.disp_edges.render)
         yield ((), 'ident', 1, self.disp_edges.identify)
     
     @property
-    def world(self):    return self.vertices.world
+    def world(self):
+        return self.vertices.world
     @world.setter
-    def world(self, value):    self.vertices.world = value
+    def world(self, value):
+        self.vertices.world = value
 
-    #def control(self, view, key, sub, evt):
-        #if evt.type() == QEvent.MouseButtonRelease and evt.button() == Qt.LeftButton:
-            #sub = sub[0]
-            #flags, idents = self.vertices.flags, self.vertices.idents
-            #for i in range(len(idents)):
-                #flags[i] ^= idents[i] == sub
-            #self.vertices.flags_updated = True
-            #view.update()
-            #evt.accept()
+    # def control(self, view, key, sub, evt):
+    #     if evt.type() == QEvent.MouseButtonRelease and evt.button() == Qt.LeftButton:
+    #         sub = sub[0]
+    #         flags, idents = self.vertices.flags, self.vertices.idents
+    #         for i in range(len(idents)):
+    #             flags[i] ^= idents[i] == sub
+    #         self.vertices.flags_updated = True
+    #         view.update()
+    #         evt.accept()
 
 
 class Vertices(object):
@@ -116,7 +127,6 @@ class Vertices(object):
         for i,id in enumerate(self.idents):
             self.flags[i] ^= id == sub
         self.flags_updated = True
-            
 
 
 class SurfaceDisplay:
